refactor(tls): refactor tomcat connector init into util class.
Allows for additional connector customizers in extension code to reuse
common TLS configuration code.
